modules order sizing, still wip
